+++ test-utils/custom-environment.ts
@@ -7,6 +7,11 @@ const matchMedia = (query: string) =>
         matches: false,
         addListener: function () {},
         removeListener: function () {},
+        addEventListener: function () {},
+        removeEventListener: function () {},
+        dispatchEvent: function () {
+            return false;
+        },
     } as unknown as MediaQueryList);
 
 class CustomEnvironment extends JsDomEnvironmetn {